Reorganize tests into own namespace
