+++ composer.json
@@ -46,7 +46,7 @@
         "ext-mailparse": "*"
     },
     "require-dev": {
-        "phpunit/phpunit":              "^8.0",
+        "phpunit/phpunit":              "^7.0",
         "phpunit/php-token-stream":     "^3.0",
         "php-coveralls/php-coveralls":  "^2.2",
         "squizlabs/php_codesniffer":    "^3.5"
@@ -57,5 +57,9 @@
     },
     "autoload": {
         "psr-4": { "PhpMimeMailParser\\": "src/" }
+    },
+    "autoload-dev": {
+        "psr-4": { "Tests\\PhpMimeMailParser\\": "tests/" },
+        "files": ["tests/globals.php"]
     }
 }
