+++ tests/AttachmentTest.php
@@ -1,9 +1,7 @@
 <?php
-namespace PhpMimeMailParser;
+namespace Tests\PhpMimeMailParser;
 
 use PhpMimeMailParser\Parser;
-use PhpMimeMailParser\Attachment;
-use PhpMimeMailParser\Exception;
 
 /**
  * Test Attachment of php-mime-mail-parser
@@ -11,7 +9,7 @@ use PhpMimeMailParser\Exception;
  * Fully Tested Mailparse Extension Wrapper for PHP 5.4+
  *
  */
-class AttachmentTest extends \PHPUnit\Framework\TestCase
+class AttachmentTest extends TestCase
 {
     public function testSaveAttachmentsFromParser()
     {
